Move twitter text filter spec to its proper place
