+++ spec/models/text_filter_spec.rb
@@ -3,32 +3,6 @@ require 'rails_helper'
 describe 'With the list of available filters', type: :model do
   let(:blog) { build_stubbed(:blog) }
 
-  describe 'Twitter filter' do
-    def filter_text(text, filters)
-      TextFilter.filter_text(text, filters)
-    end
-
-    it 'should replace a hashtag with a proper URL to Twitter search' do
-      text = filter_text('A test tweet with a #hashtag', [:twitterfilter])
-      expect(text).to eq("A test tweet with a <a href='https://twitter.com/search?q=%23hashtag&src=tren&mode=realtime'>#hashtag</a>")
-    end
-
-    it 'should replace a @mention by a proper URL to the twitter account' do
-      text = filter_text('A test tweet with a @mention', [:twitterfilter])
-      expect(text).to eq("A test tweet with a <a href='https://twitter.com/mention'>@mention</a>")
-    end
-
-    it 'should replace a http URL by a proper link' do
-      text = filter_text('A test tweet with a http://link.com', [:twitterfilter])
-      expect(text).to eq("A test tweet with a <a href='http://link.com'>http://link.com</a>")
-    end
-
-    it 'should replace a https URL with a proper link' do
-      text = filter_text('A test tweet with a https://link.com', [:twitterfilter])
-      expect(text).to eq("A test tweet with a <a href='https://link.com'>https://link.com</a>")
-    end
-  end
-
   describe '#filter_text' do
     def filter_text(text, filters)
       TextFilter.filter_text(text, filters)
@@ -37,11 +11,6 @@ describe 'With the list of available filters', type: :model do
     it 'unknown' do
       text = filter_text('*foo*', [:unknowndoesnotexist])
       expect(text).to eq('*foo*')
-    end
-
-    it 'Twitter' do
-      text = filter_text('A test tweet with a #hashtag and a @mention', [:twitterfilter])
-      expect(text).to eq("A test tweet with a <a href='https://twitter.com/search?q=%23hashtag&src=tren&mode=realtime'>#hashtag</a> and a <a href='https://twitter.com/mention'>@mention</a>")
     end
 
     it 'smartypants' do
